Allow null to be passed as libpath. The result is that a series of procedures for locating the libnode library is done. a) On Core+ platforms, NativeLibrary is used directly with a relative path off of the assembly. This causes a search across the applications .deps.json. This is the ideal scenario as long as the LibNode packages deliver the library into runtimes/{rid}/native. b) On Framework platforms a search procedure is used. First, libnode is attempted to be loaded by relative name directly. This handles the situation where the library is directly available in some way. Such as relative to the executable. Second runtimes/{rid}/native is examined with a known RID based on current platform.
A few local NativeLibrary methods were improved.

Load was updated to support non-Windows platforms. This handles the scenario of Framework on ~Windows. Such as Mono.
TryLoad was introduced to probe paths during discovery.
GetExport/TryGetExport was updated to support non-Windows platforms. This handles the scenario of Framework on ~Windows. Such as Mono.
dlerror() was introduced to handle the situation of null exports. Which should not reaaaaly happen in libnode. But is technically correct.
The pinvokes to libdl are technically out of date. dl was merged into libc on Linux. And was already present in libc on OS X. But these are left.
I made changes to NuGet.config, removing the react stuff and the disabled sources. I don't know what the goal was here, but this was required to get the code building for me.
Expects to use Microsoft.JavaScript.LibNode version that deposit native libraries properly.
Change tests to not need a hard coded libnode path.